requirements.txt
crispy-tailwind==1.0.3
Django==5.0.3
django-crispy-forms==2.1
djangorestframework==3.15.1
pillow==10.3.0
stripe==9.1.0
Unidecode==1.3.8

